✨ feat: try split entries for coin-ethereum for better treeshaking
