+++ packages/coin-ethereum/package.json
@@ -6,18 +6,28 @@
   "files": [
     "dist"
   ],
-  "main": "./dist/coin-ethereum.umd.cjs",
-  "module": "./dist/coin-ethereum.js",
+  "main": "./dist/index.umd.cjs",
+  "module": "./dist/index.js",
   "types": "dist/index.d.ts",
   "exports": {
     ".": {
       "import": {
         "types": "./dist/index.d.ts",
-        "default": "./dist/coin-ethereum.js"
+        "default": "./dist/index.js"
       },
       "require": {
         "types": "./dist/index.d.ts",
-        "default": "./dist/coin-ethereum.umd.cjs"
+        "default": "./dist/index.umd.cjs"
+      }
+    },
+    "./chains": {
+      "import": {
+        "types": "./dist/chains.d.ts",
+        "default": "./dist/chains.js"
+      },
+      "require": {
+        "types": "./dist/chains.d.ts",
+        "default": "./dist/chains.umd.cjs"
       }
     }
   },
