+++ packages/coin-ethereum/src/chain-wallet/index.ts
@@ -1,2 +1,1 @@
-export * from './chain'
 export * from './wallet'
