+++ packages/coin-ethereum/src/chains.ts
